+++ src/core/fields/mod.rs
@@ -10,6 +10,7 @@ use super::backend::ColumnOps;
 pub mod cm31;
 pub mod m31;
 pub mod qm31;
+pub mod secure_column;
 
 pub trait FieldOps<F: Field>: ColumnOps<F> {
     // TODO(Ohad): change to use a mutable slice.
