+++ src/core/poly/circle/mod.rs
@@ -10,7 +10,7 @@ pub use domain::{CircleDomain, MAX_CIRCLE_DOMAIN_LOG_SIZE};
 pub use evaluation::{CircleEvaluation, CosetSubEvaluation};
 pub use ops::PolyOps;
 pub use poly::CirclePoly;
-pub use secure_poly::{combine_secure_value, SecureCirclePoly};
+pub use secure_poly::SecureCirclePoly;
 
 #[cfg(test)]
 mod tests {
