+++ src/core/poly/circle/secure_poly.rs
@@ -1,15 +1,15 @@
 use std::ops::Deref;
 
-use crate::core::air::evaluation::SECURE_EXTENSION_DEGREE;
 use crate::core::backend::cpu::CPUCirclePoly;
 use crate::core::circle::CirclePoint;
 use crate::core::fields::qm31::SecureField;
+use crate::core::fields::secure_column::SECURE_EXTENSION_DEGREE;
 
 pub struct SecureCirclePoly(pub [CPUCirclePoly; SECURE_EXTENSION_DEGREE]);
 
 impl SecureCirclePoly {
     pub fn eval_at_point(&self, point: CirclePoint<SecureField>) -> SecureField {
-        combine_secure_value(self.eval_columns_at_point(point))
+        Self::eval_from_partial_evals(self.eval_columns_at_point(point))
     }
 
     pub fn eval_columns_at_point(
@@ -23,6 +23,16 @@ impl SecureCirclePoly {
             self[3].eval_at_point(point),
         ]
     }
+
+    /// Evaluates the polynomial at a point, given evaluations of its composing base field
+    /// polynomials at that point.
+    pub fn eval_from_partial_evals(evals: [SecureField; SECURE_EXTENSION_DEGREE]) -> SecureField {
+        let mut res = evals[0];
+        res += evals[1] * SecureField::from_u32_unchecked(0, 1, 0, 0);
+        res += evals[2] * SecureField::from_u32_unchecked(0, 0, 1, 0);
+        res += evals[3] * SecureField::from_u32_unchecked(0, 0, 0, 1);
+        res
+    }
 }
 
 impl Deref for SecureCirclePoly {
@@ -31,12 +41,4 @@ impl Deref for SecureCirclePoly {
     fn deref(&self) -> &Self::Target {
         &self.0
     }
-}
-
-pub fn combine_secure_value(value: [SecureField; SECURE_EXTENSION_DEGREE]) -> SecureField {
-    let mut res = value[0];
-    res += value[1] * SecureField::from_u32_unchecked(0, 1, 0, 0);
-    res += value[2] * SecureField::from_u32_unchecked(0, 0, 1, 0);
-    res += value[3] * SecureField::from_u32_unchecked(0, 0, 0, 1);
-    res
 }
